+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Fix warm start detection ([#3937](https://github.com/getsentry/sentry-java/pull/3937))
 
+### Dependencies
+
+- Bump Native SDK from v0.7.16 to v0.7.17 ([#4003](https://github.com/getsentry/sentry-java/pull/4003))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0717)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.7.16...0.7.17)
+
 ## 7.19.1
 
 ### Fixes