+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-### Unreleased
+## Unreleased
+
+### Fixes
+
+- Fix warm start detection ([#3937](https://github.com/getsentry/sentry-java/pull/3937))
+
+## 7.19.1
 
 ### Fixes